Fix unordered map successfully, reason: .get return Option<T>, sol: reassign

+++ src/types.rs
@@ -1,5 +1,5 @@
-pub type BaggageId = usize;
-pub type FlightId = usize;
+pub type BaggageId = u64;
+pub type FlightId = u64;
 pub type Distance = f32;    // miles    
 pub type Price = u128;  // yoto NEAR
 pub type Weight = f32;  // pound (lbs)
